move join tests
